requirements.txt
# BTC Intelligence Dependencies

# WebSocket and HTTP
websockets>=11.0
certifi>=2023.0.0
aiohttp>=3.8.0
aiohttp-cors>=0.7.0

# Web UI
flask>=2.3.0
flask-socketio>=5.3.0

# Privy token auto-refresh
playwright>=1.40.0

# Hyperliquid SDK (for REST API)
hyperliquid-python-sdk>=0.8.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Data analysis
numpy>=1.24.0
pandas>=2.0.0

# Telegram notifications (optional)
python-telegram-bot>=20.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
